+++ lib/pigeon/apns_config.ex
@@ -14,7 +14,8 @@ defmodule Pigeon.APNS.Config do
       certfile: file_path(config[:cert]),
       key: key(config[:key]),
       keyfile: file_path(config[:key]),
-      use_2197: config[:use_2197] || false
+      use_2197: config[:use_2197] || false,
+      ping_period: config[:ping_period] || 600_000
     }
   end