+++ lib/pigeon/apns_worker.ex
@@ -1,11 +1,9 @@
 defmodule Pigeon.APNSWorker do
   @moduledoc """
-    Handles all APNS request and response parsing over an HTTP2 connection.
+  Handles all APNS request and response parsing over an HTTP2 connection.
   """
   use GenServer
   require Logger
-
-  @ping_period 600_000 # 10 minutes
 
   defp apns_production_api_uri, do: "api.push.apple.com"
   defp apns_development_api_uri, do: "api.development.push.apple.com"
@@ -18,7 +16,10 @@ defmodule Pigeon.APNSWorker do
   end
 
   def start_link(config) do
-    GenServer.start_link(__MODULE__, {:ok, config}, name: config[:name])
+    case config[:name] do
+      nil -> GenServer.start_link(__MODULE__, {:ok, config})
+      name -> GenServer.start_link(__MODULE__, {:ok, config}, name: name)
+    end
   end
 
   def stop, do: :gen_server.cast(self(), :stop)
@@ -29,7 +30,7 @@ defmodule Pigeon.APNSWorker do
     mode = config[:mode]
     case connect_socket(config, 0) do
       {:ok, socket} ->
-        Process.send_after(self(), :ping, @ping_period)
+        Process.send_after(self(), :ping, config[:ping_period])
         {:ok, %{
           apns_socket: socket,
           mode: mode,
@@ -100,7 +101,7 @@ defmodule Pigeon.APNSWorker do
   end
 
   defp do_connect_socket(config, uri, options, tries) do
-    case Kadabra.open(uri, :https, options) do
+    case Pigeon.Http2.Client.default().connect(uri, :https, options) do
       {:ok, socket} -> {:ok, socket}
       {:error, reason} ->
         Logger.error(inspect(reason))
@@ -133,10 +134,10 @@ defmodule Pigeon.APNSWorker do
       |> put_apns_id(notification)
       |> put_apns_topic(notification)
 
-    Kadabra.request(socket, req_headers, json)
+    Pigeon.Http2.Client.default().send_request(socket, req_headers, json)
     new_q = Map.put(queue, "#{stream_id}", {notification, on_response})
     new_stream_id = stream_id + 2
-    { :noreply, %{state | stream_id: new_stream_id, queue: new_q } }
+    {:noreply, %{state | stream_id: new_stream_id, queue: new_q}}
   end
 
   defp put_apns_id(headers, notification) do
@@ -216,7 +217,7 @@ defmodule Pigeon.APNSWorker do
       :missing_provider_token ->
         """
         No provider certificate was used to connect to APNs and Authorization header was missing
-        or no provider token was specified." 
+        or no provider token was specified."
         """
 
       # 404
@@ -259,14 +260,21 @@ defmodule Pigeon.APNSWorker do
   end
 
   def handle_info(:ping, state) do
-    Kadabra.ping(state.apns_socket)
-    Process.send_after(self(), :ping, @ping_period)
+    Pigeon.Http2.Client.default().send_ping(state.apns_socket)
+    Process.send_after(self(), :ping, state.config.ping_period)
 
-    { :noreply, state }
+    {:noreply, state}
   end
 
-  def handle_info({:end_stream, %Kadabra.Stream{id: stream_id, headers: headers, body: body}},
-                                                %{apns_socket: _socket, queue: queue} = state) do
+  def handle_info(msg, state) do
+    case Pigeon.Http2.Client.default().handle_end_stream(msg, state) do
+      {:ok, %Pigeon.Http2.Stream{} = stream} -> process_end_stream(stream, state)
+      _else -> {:noreply, state}
+    end
+  end
+
+  def process_end_stream(%Pigeon.Http2.Stream{id: stream_id, headers: headers, body: body},
+                                            %{apns_socket: _socket, queue: queue} = state) do
 
     {notification, on_response} = queue["#{stream_id}"]
 
@@ -286,10 +294,6 @@ defmodule Pigeon.APNSWorker do
         {:noreply, %{state | queue: new_queue}}
     end
   end
-
-  def handle_info({:ping, _from}, state), do: {:noreply, state}
-
-  def handle_info({:ok, _from}, state), do: {:noreply, state}
 
   defp get_status(headers) do
     case Enum.find(headers, fn({key, _val}) -> key == ":status" end) do