+++ lib/pigeon/gcm.ex
@@ -3,150 +3,138 @@ defmodule Pigeon.GCM do
   Handles all Google Cloud Messaging (GCM) request and response functionality.
   """
   require Logger
+  import Supervisor.Spec
 
-  # defp gcm_uri, do: 'https://gcm-http.googleapis.com/gcm/send'
-  defp gcm_uri, do: 'fcm.googleapis.com/fcm/send'
+  alias Pigeon.GCM.NotificationResponse
+  alias Pigeon.GCM.Notification
 
-  defp gcm_headers(key) do
-    [{ "Authorization", "key=#{key}" },
-     { "Content-Type", "application/json" },
-     { "Accept", "application/json" }]
+  @default_timeout 5_000
+
+  def push(notification, opts \\ [])
+  def push(notification, opts) when is_list(notification) do
+    case opts[:on_response] do
+      nil ->
+        ref = :erlang.make_ref
+        pid = self()
+        for n <- notification do
+          on_response = fn(x) -> send pid, {ref, x} end
+          send_push(n, on_response, opts)
+        end
+        Enum.foldl(notification, %{}, fn(n, acc) ->
+          receive do
+            {^ref, %NotificationResponse{message_id: id} = response} ->
+              if Map.has_key?(acc, id) do
+                %{acc | id => merge(response, acc[:message_id])}
+              else
+                Map.merge(%{id => response}, acc)
+              end
+          after 5_000 ->
+            acc
+          end
+        end)
+      on_response -> send_push(notification, on_response, opts)
+    end
   end
 
-  defp default_gcm_key, do: Application.get_env(:pigeon, :gcm)[:key]
-
-  @doc """
-  Sends a push over GCM
-  """
-  @spec push(Pigeon.GCM.Notification) :: none
-  def push(notification) do
-    do_push(notification, %{gcm_key: default_gcm_key()})
+  def push(notification, opts) do
+    case opts[:on_response] do
+      nil -> do_sync_push(notification, opts)
+      on_response -> send_push(notification, on_response, opts)
+    end
   end
 
-  @doc """
-  Sends a push over GCM and executes function on success/failure.
-  """
-  @spec push(Pigeon.GCM.Notification, (() -> none)) :: none
-  def push(notification, on_response) when is_function(on_response) do
-    do_push(notification, %{gcm_key: default_gcm_key()}, on_response)
+  defp do_sync_push(notification, opts) do
+    ref = :erlang.make_ref
+    pid = self()
+    on_response = fn(x) -> send pid, {ref, x} end
+    send_push(notification, on_response, opts)
+    receive do
+      {^ref, x} -> x
+    after
+      @default_timeout -> {:error, :timeout, notification}
+    end
   end
 
-  def push(notification, config, on_response \\ nil) do
-    do_push(notification, config, on_response)
+  def encode_requests(%{registration_id: regid} = notification) when is_binary(regid) do
+    encode_requests(%{notification | registration_id: [regid]})
+  end
+  def encode_requests(%{registration_id: regid} = notification) when length(regid) < 1001 do
+      res =
+        regid
+        |> recipient_attr()
+        |> Map.merge(notification.payload)
+        |> Poison.encode!
+        formatted_regid = regid
+        |> List.wrap
+
+      [{formatted_regid, res}]
   end
 
-  defp do_push(notification, %{gcm_key: gcm_key}, on_response \\ nil) do
-    requests =
+  def encode_requests(notification) do
       notification.registration_id
-      |> chunk_registration_ids
-      |> encode_requests(notification)
+      |> Enum.chunk(1000, 1000, [])
+      |> Enum.map(& encode_requests(%{notification | registration_id: &1}))
+      |> List.flatten
+  end
 
-    response =
-      case on_response do
-        nil ->
-          fn({_reg_ids, payload}) ->
-            HTTPoison.post(gcm_uri(), payload, gcm_headers(gcm_key))
-          end
-        _ ->
-          fn({reg_ids, payload}) ->
-            case HTTPoison.post(gcm_uri(), payload, gcm_headers(gcm_key)) do
-              {:ok, %HTTPoison.Response{status_code: status, body: body}} ->
-                notification = %{ notification | registration_id: reg_ids }
-                process_response(status, body, notification, on_response)
-              _ ->
-                process_response(:unknown, "", notification, on_response)
-            end
+  defp recipient_attr([regid]), do: %{"to" => regid}
+  defp recipient_attr(regid) when is_list(regid), do: %{"registration_ids" => regid}
 
-          end
+  @doc """
+    Sends a push over GCM.
+  """
+  def send_push(notification, on_response, opts) do
+    notification
+    |> encode_requests()
+    |> Enum.map(& GenServer.cast(:gcm_worker, generate_envelope(&1, on_response, opts)))
+  end
+
+  def start_connection(name) do
+    config = %{
+      name: name,
+      gcm_key:  Application.get_env(:pigeon, :gcm)[:key]
+    }
+    Supervisor.start_child(:pigeon, worker(Pigeon.GCMWorker, [config], id: name))
+  end
+
+  def stop_connection(name) do
+    Supervisor.terminate_child(:pigeon, name)
+    Supervisor.delete_child(:pigeon, name)
+  end
+
+  def generate_envelope(payload, on_response, opts) do
+    {:push, :gcm, payload, on_response, Map.new(opts)}
+  end
+
+  # def merge(%NotificationResponse{ok: ok1,
+  #                                 retry: retry1,
+  #                                 update: update1,
+  #                                 remove: remove1,
+  #                                 error: error1}, %NotificationResponse{ok: ok2,
+  #                                                                       retry: retry2,
+  #                                                                       update: update2,
+  #                                                                       remove: remove2,
+  #                                                                       error: error2}) do
+
+  #   error3 = Map.merge(error1, error2, fn(m, a, b) ->  a ++ b end)
+  #   %NotificationResponse{
+  #     ok: ok1 ++ ok2,
+  #     retry: retry1 ++ retry2,
+  #     update: update1 ++ update2,
+  #     remove: remove1 ++ remove2,
+  #     error: error3
+  #   }
+  # end
+
+  def merge(response_1, response_2) do
+    Map.merge(response_1, response_2, fn(key, value_1, value_2) ->
+      cond do
+        key == :__struct__ -> value_1
+        is_map(value_1) -> merge(value_1, value_2)
+        is_nil(value_1) -> value_2
+        is_nil(value_2) -> value_1
+        true -> value_1 ++ value_2
       end
-    for r <- requests, do: Task.async(fn -> response.(r) end)
-    :ok
-  end
-
-  def chunk_registration_ids(reg_ids) when is_binary(reg_ids), do: [[reg_ids]]
-  def chunk_registration_ids(reg_ids), do: Enum.chunk(reg_ids, 1000, 1000, [])
-
-  def encode_requests([[reg_id]|_rest], %{payload: payload, priority: priority}) do
-    to_send = Map.merge(%{"to" => reg_id, "priority" => to_string(priority)}, payload)
-    [{reg_id, Poison.encode!(to_send)}]
-  end
-  def encode_requests(registration_ids, %{payload: payload, priority: priority}) do
-    Enum.map(registration_ids, fn(x) -> encode_payload(x, priority, payload) end)
-  end
-
-  defp encode_payload(x, priority, payload) do
-    encoded =
-      %{"registration_ids" => x, "priority" => to_string(priority)}
-      |> Map.merge(payload)
-      |> Poison.encode!
-    {x, encoded}
-  end
-
-  def process_response(status, body, notification, on_response) do
-    case status do
-      200 ->
-        handle_200_status(body, notification, on_response)
-      400 ->
-        handle_error_status_code(:invalid_json, notification, on_response)
-      401 ->
-        handle_error_status_code(:authentication_error, notification, on_response)
-      500 ->
-        handle_error_status_code(:internal_server_error, notification, on_response)
-      _ ->
-        handle_error_status_code(:unknown_error, notification, on_response)
-    end
-  end
-
-  def handle_error_status_code(reason, notification, on_response),
-    do: on_response.({:error, reason, notification})
-
-  def handle_200_status(body, %{registration_id: reg_id} = n, on_response) when is_list(reg_id) do
-    {:ok, json} = Poison.decode(body)
-    results = Enum.zip(n.registration_id, json["results"])
-    for result <- results, do: process_callback(result, n, on_response)
-  end
-  def handle_200_status(body, %{registration_id: _reg_id} = notification, on_response) do
-    {:ok, json} = Poison.decode(body)
-    results = Enum.zip([notification.registration_id], json["results"])
-    for result <- results, do: process_callback(result, notification, on_response)
-  end
-
-  def process_callback({reg_id, response}, notification, on_response) do
-    case parse_result(response) do
-      {:ok, message_id} ->
-        notification = %{ notification | registration_id: reg_id, message_id: message_id }
-        on_response.({:ok, notification})
-
-      {:ok, message_id, registration_id} ->
-        notification =
-          %{ notification | registration_id: reg_id,
-          message_id: message_id,
-          updated_registration_id: registration_id }
-        on_response.({:ok, notification})
-
-      {:error, reason} ->
-        notification = %{ notification | registration_id: reg_id }
-        on_response.({:error, reason, notification})
-    end
-  end
-
-  def parse_result(result) do
-    error = result["error"]
-    if is_nil(error) do
-      parse_success(result)
-    else
-      error_atom = error |> Macro.underscore |> String.to_atom
-      {:error, error_atom}
-    end
-  end
-
-  def parse_success(result) do
-    message_id = result["message_id"]
-    registration_id = result["registration_id"]
-    if is_nil(registration_id) do
-      {:ok, message_id}
-    else
-      {:ok, message_id, registration_id}
-    end
+    end)
   end
 end